Updated the omssa parser to version 1.6.3, fixing a problem with parsing omx files with intensity scale was a floating value. Updated X!Tandem parser to version 1.7.7. Updated ols-dialog to version 3.4.2. Moved project to version 2.5.7-beta.

+++ trunk/pride_converter/prideconverter-core/src/main/java/no/uib/prideconverter/OmssaConverter.java
@@ -76,7 +76,7 @@ public class OmssaConverter {
         ArrayList<CvParam> cVParams;
         ArrayList<UserParam> userParams;
         List<Integer> mzValues, intensityValues;
-        long omssaAbundanceScale;
+        double omssaAbundanceScale;
         Collection<SpectrumDescComment> spectrumDescriptionComments;
 
         MSModHit currentMSModHit;
